% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{analyze_sentiment}
\alias{analyze_sentiment}
\title{analyze_sentiment}
\usage{
analyze_sentiment(data)
}
\arguments{
\item{data}{Vector of plain text to analyze.}
}
\value{
List of parsed response objects.
}
\description{
Returns sentiment analyzis from Google cloud language API.
}
