% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{as.data.frame.sentiment_analysis}
\alias{as.data.frame.sentiment_analysis}
\title{sentiment analysis data frame}
\usage{
\method{as.data.frame}{sentiment_analysis}(x)
}
\arguments{
\item{data}{Data of class sentiment_analysis}
}
\value{
Data frame
}
\description{
sentiment analysis data frame
}
