% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{as_tibble.sentiment_analysis}
\alias{as_tibble.sentiment_analysis}
\title{sentiment analysis tbl}
\usage{
as_tibble.sentiment_analysis(data)
}
\arguments{
\item{data}{Data of class sentiment_analysis}
}
\value{
A tbl
}
\description{
sentiment analysis tbl
}
