% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\docType{data}
\name{inaug45}
\alias{inaug45}
\title{inaug45}
\format{A character vector consisting of 75 paragraphs.}
\source{
https://www.whitehouse.gov/inaugural-address
}
\usage{
inaug45
}
\description{
Transcript of Trump's 2017 Inaugural Address tokenized by paragraph.
}
\keyword{datasets}
