% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-call.R
\name{update_api_base_url}
\alias{update_api_base_url}
\title{update base url}
\usage{
update_api_base_url(scheme, base, version)
}
\arguments{
\item{scheme}{http or https}

\item{base}{base api, e.g., api.twitter.com}

\item{version}{version string, e.g., v1.1}
}
\description{
update base url
}
