% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{analyze_sentiment}
\alias{analyze_sentiment}
\alias{analyse_sentiment}
\title{analyze_sentiment}
\usage{
analyze_sentiment(text, id = NULL)
}
\arguments{
\item{text}{Vector of plain text to analyze.}
}
\value{
List of parsed response objects.
}
\description{
Conducts and returns results of sentiment analysis using Google Cloud Natural
  Language API.
}
